몬스터 길찾기 오류 수정

+++ ServerPortfolio/ZoneServer/ZoneServer/Monster.cpp
@@ -33,19 +33,19 @@ Monster::Monster(Field* _field, FieldTilesData* _fieldTilesData,
 	m_sectorManager = _sectorManager;
 	m_sector = nullptr;
 
-	Spawn();
-
 	InitializeCriticalSection(&m_cs);
+
+	Spawn();
 };
 
 Monster::~Monster()
 {
 	if(m_sendBuffer != nullptr) delete m_sendBuffer;
 
+	DeleteCriticalSection(&m_cs);
+
 	m_tileList.clear();
 	m_tileList.resize(0);
-
-	DeleteCriticalSection(&m_cs);
 }
 
 void Monster::Reset()
@@ -107,9 +107,9 @@ void Monster::Spawn()
 
 	m_sendBuffer->Write(monsterInfoPacket->size);
 
-	MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterInfoPacket });
-	/*m_field->SectorSendAll(m_sector->GetRoundSectorsVec(),
-		reinterpret_cast<char*>(monsterInfoPacket), monsterInfoPacket->size);*/
+	//MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterInfoPacket });
+	m_field->SectorSendAll(m_sector->GetRoundSectorsVec(),
+		reinterpret_cast<char*>(monsterInfoPacket), monsterInfoPacket->size);
 }
 
 void Monster::Move()
@@ -144,8 +144,8 @@ void Monster::Attack()
 
 		//MYDEBUG("따라감 \n");
 
-		PathFindStart(m_targetTile);
-		m_info.state = STATE::MOVE;
+		PathFindStart(m_targetTile, STATE::MOVE);
+		//m_info.state = STATE::MOVE;
 
 		return;
 	}
@@ -156,8 +156,8 @@ void Monster::Attack()
 		//MYDEBUG("죽었으니 돌아가기 \n");
 
 		m_targetTile = m_homeTile;
-		PathFindStart(m_targetTile);
-		m_info.state = STATE::RETURN;
+		PathFindStart(m_targetTile, STATE::RETURN);
+		//m_info.state = STATE::RETURN;
 
 		return;
 	}
@@ -173,8 +173,7 @@ void Monster::Attack()
 		monsterAttackPacket->Init(SendCommand::Zone2C_MONSTER_ATTACK, sizeof(MonsterAttackPacket));
 
 		m_sendBuffer->Write(monsterAttackPacket->size);
-		MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterAttackPacket });
-		
+		//MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterAttackPacket });
 		/*m_field->SectorSendAll(m_sector->GetRoundSectorsVec(),
 			reinterpret_cast<char*>(monsterAttackPacket), monsterAttackPacket->size);*/
 
@@ -186,8 +185,6 @@ void Monster::Attack()
 
 Packet* Monster::Hit(User* _user, int _damage)
 {
-	CSLock cslock(m_cs);
-
 	if (m_info.state == STATE::DEATH) return nullptr;
 
 	//대부분 알피지는 이럴때 공격이 안됐음
@@ -253,8 +250,17 @@ void Monster::FSM()
 					GetTile(m_homeTile->GetX() + randX, m_homeTile->GetY() + randY);
 			}
 
-			PathFindStart(m_targetTile);
-			m_info.state = STATE::PATROL;
+			PathFindStart(m_targetTile, STATE::PATROL);
+			break;
+		}
+		case STATE::READY:
+		{
+			if (m_currentTime <= 10.0f) break;
+
+			MYDEBUG("[ %d monster : ReSend ] \n", m_info.index);
+			m_targetTile = m_homeTile;
+			PathFindStart(m_targetTile, STATE::RETURN);
+
 			break;
 		}
 		//정찰
@@ -291,8 +297,8 @@ void Monster::FSM()
 				//MYDEBUG("멀어서 귀환 \n");
 
 				m_targetTile = m_homeTile;
-				PathFindStart(m_targetTile);
-				m_info.state = STATE::RETURN;
+				PathFindStart(m_targetTile, STATE::RETURN);
+				//m_info.state = STATE::RETURN;
 
 				break;
 			}
@@ -301,7 +307,7 @@ void Monster::FSM()
 			if (m_targetTile != m_target->GetTile())
 			{
 				m_targetTile = m_target->GetTile();
-				PathFindStart(m_targetTile);
+				PathFindStart(m_targetTile, STATE::MOVE);
 
 				break;
 			}
@@ -360,69 +366,72 @@ void Monster::FSM()
 	}
 }
 
-void Monster::PathFindStart(Tile* _targetTile)
+void Monster::PathFindStart(Tile* _targetTile, STATE _state)
 {
-	/*m_currentTime = 0.0f;
+	//MYDEBUG("[ %d monster : Start ] \n", m_info.index);
+
+	char sendBuffer[1000];
 
 	m_info.state = STATE::READY;
+	m_currentTime = 0;
 
-	m_tileList.clear();
-
-	PathFindPacket* pathFindPacket =
-		reinterpret_cast<PathFindPacket*>(m_sendBuffer->GetBuffer(sizeof(PathFindPacket)));
+	PathFindPacket* pathFindPacket = reinterpret_cast<PathFindPacket*>(sendBuffer);
+		//reinterpret_cast<PathFindPacket*>(m_sendBuffer->GetBuffer(sizeof(PathFindPacket)));
 	pathFindPacket->Init(SendCommand::Zone2Path_PATHFIND, sizeof(PathFindPacket));
 
-	pathFindPacket->state = _state;
-	pathFindPacket->nowTile = m_nowTile;
-	pathFindPacket->targetTile = m_targetTile;*/
-
-	/*pathFindPacket->monsterNum = m_info.index;
+	pathFindPacket->monsterNum = m_info.index;
 	pathFindPacket->fieldNum = m_field->GetFieldNum();
 	pathFindPacket->state = _state;
 
 	pathFindPacket->nowPosition.x = m_nowTile->GetX();
 	pathFindPacket->nowPosition.y = m_nowTile->GetY();
 	pathFindPacket->targetPosition.x = _targetTile->GetX();
-	pathFindPacket->targetPosition.y = _targetTile->GetY();*/
+	pathFindPacket->targetPosition.y = _targetTile->GetY();
 
-	//MainThread::getSingleton()->AddToPathFindQueue({ this, pathFindPacket });
-	
-	/*PathFinderAgent::getSingleton()->Send(reinterpret_cast<char*>(pathFindPacket), 
-		pathFindPacket->size);*/
+	//m_sendBuffer->Write(pathFindPacket->size);
+
+	PathFinderAgent::getSingleton()->
+		Send(reinterpret_cast<char*>(pathFindPacket), pathFindPacket->size);
+
+	m_info.state = STATE::IDLE;
 
 	/*PathFinderAgent::getSingleton()->SendPathFindPacket(m_info.index, m_field->GetFieldNum(), _state,
 		m_nowTile->GetX(), m_nowTile->GetY(), _targetTile->GetX(), _targetTile->GetY());*/
 
-	PathFinding::PathFind(&m_tileList, m_nowTile, _targetTile);
+	/*PathFinding::PathFind(&m_tileList, m_nowTile, _targetTile);
 
-	//MYDEBUG("target : %d , %d \n", _targetTile->GetX(), _targetTile->GetY());
+	MYDEBUG("target : %d , %d \n", _targetTile->GetX(), _targetTile->GetY());
 
 	if (m_tileList.size() == 0) return;
 
 	m_targetPosition = m_tileList.front();
 
 	m_tileList.pop_front();
-	m_currentTime = 0;
+	m_currentTime = 0;*/
 }
 
-void Monster::PathFindSuccess(unsigned short _state)
+void Monster::PathFindSuccess(PathFindPacket_Success* _packet)
 {
-	//MYDEBUG("[ %d monster Success ] \n", m_info.index);
+	//MYDEBUG("[ %d monster : Success ] \n", m_info.index);
+
+	for (int i = 0; i < _packet->listCount; i++)
+	{
+		m_tileList.push_back(_packet->positionList[i]);
+	}
 
 	m_targetPosition = m_tileList.front();
 
 	m_tileList.pop_front();
 
-	m_info.state = (STATE)_state;
+	m_info.state = (STATE)_packet->state;
 	m_currentTime = 0;
 }
 
 void Monster::PathFindFailed()
 {
-	//MYDEBUG("[ %d monster Failed ] \n", m_info.index);
+	//MYDEBUG("[ %d monster : Failed ] \n", m_info.index);
 
-	//m_targetTile = m_homeTile;
-	//PathFindStart(m_targetTile, STATE::RETURN);
+	PathFindStart(m_targetTile, STATE::RETURN);
 }
 
 bool Monster::PathMove()
@@ -458,9 +467,9 @@ bool Monster::PathMove()
 
 			//MYDEBUG("my position : %f , %f \n", m_info.position.x, m_info.position.y);
 
-			MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterPositionPacket });
-			/*m_field->SectorSendAll(m_sector->GetRoundSectorsVec(),
-				reinterpret_cast<char*>(monsterPositionPacket), monsterPositionPacket->size);*/
+			//MainThread::getSingleton()->AddToMonsterPacketQueue({ this, monsterPositionPacket });
+			m_field->SectorSendAll(m_sector->GetRoundSectorsVec(),
+				reinterpret_cast<char*>(monsterPositionPacket), monsterPositionPacket->size);
 		}
 		//남은 타일리스트가 없다면
 		else
@@ -488,15 +497,23 @@ void Monster::UpdateSector()
 		Sector* nowSector = m_sectorManager->
 			GetSector(m_nowTile->GetX(), m_nowTile->GetY());
 
-		Sector* prevSector = m_sector;
-
-		/*if (prevSector != m_sector)
+		//현재 섹터와 저장되어있던 나의 섹터가 다름
+		if (nowSector != m_sector)
 		{
+			Sector* prevSector = m_sector;
 			if (prevSector != nullptr)
 			{
+				//printf("[ Exit Monster (Prev Sector) ]");
 				prevSector->GetMonsterList()->DeleteItem(this);
 			}
-		}*/
+
+			m_sector = nowSector;
+			m_sector->GetMonsterList()->AddItem(this);
+		}
+
+		//같으면 처리 X
+
+		/*Sector* prevSector = m_sector;
 
 		//현재 섹터와 저장되어있던 나의 섹터가 같음
 		if (prevSector == nowSector)
@@ -505,17 +522,12 @@ void Monster::UpdateSector()
 			return;
 		}
 
-		//if (prevSector != nullptr)
-		//{
-		//	//printf("[ Exit Monster (Prev Sector) ]");
+		if (prevSector != nullptr)
+		{
+			//printf("[ Exit Monster (Prev Sector) ]");
+		}
 
-		//	prevSector->GetMonsterList()->DeleteItem(this);
-
-		//	m_sector = nowSector;
-		//	m_sector->GetMonsterList()->AddItem(this);
-		//}
-
-		/*m_sector = nowSector;
+		m_sector = nowSector;
 		m_sector->GetMonsterList()->AddItem(this);*/
 	}
 }
